requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
google-genai
google-cloud-aiplatform
google-adk
redis
python-multipart
httpx
httpx-sse
pydantic-settings
pydantic
gspread
idna
packaging
pyasn1
pyasn1-modules
pyparsing
typing-extensions
charset-normalizer
certifi
urllib3
h11
httpcore
anyio
starlette